Graph support via. PChart
