+++ extensions/pchart/lib/config/generators.php
@@ -1,0 +1,13 @@
+<?php
+
+return [
+    'pchart' => [
+        'generator' => 'pchart',
+        'iterations' => true,
+        'height' => 800,
+        'col' => 'time',
+        'break' => [
+            'benchmark', 'subject', 'params'
+        ],
+    ],
+];
