+++ lib/Report/schema/report.xsd
@@ -14,6 +14,7 @@
         <xsd:choice maxOccurs="unbounded" minOccurs="0">
             <xsd:element name="description" type="xsd:string"/>
             <xsd:element name="table" type="table"/>
+            <xsd:element name="image" type="image"/>
         </xsd:choice>
         <xsd:attribute name="title" type="xsd:string"/>
     </xsd:complexType>
@@ -23,6 +24,10 @@
             <xsd:any processContents="skip"/>
         </xsd:choice>
         <xsd:attribute name="title" type="xsd:string"/>
+    </xsd:complexType>
+
+    <xsd:complexType name="image">
+        <xsd:attribute name="path" type="xsd:string"/>
     </xsd:complexType>
 
 </xsd:schema>
